requirements.txt
streamlit
streamlit-webrtc
langchain-core
langchain-openai
tldextract
pyzbar
pillow